backend/requirements.txt
# --- Core Frameworks ---
Flask==2.3.3
flask-cors==4.0.0

# --- Machine Learning / Deep Learning ---
tensorflow==2.13.0
scikit-learn==1.3.0
joblib==1.3.2

# --- Data Handling ---
numpy==1.24.3
pandas==2.0.3

# --- Image Processing ---
opencv-python==4.8.1.78
Pillow==10.0.1
albumentations==1.4.14

# --- Visualization ---
matplotlib==3.9.2

# --- APIs & Networking ---
requests==2.31.0
earthengine-api==1.4.3
google-api-python-client==2.154.0

# --- Dataset Management ---
kaggle==1.6.17
